pyproject.toml
[project]
name = "backtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.1",
    "ruff>=0.12.9",
    "ty>=0.0.1a19",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "polygon-api-client>=1.15.3",
]
